requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
sanic==22.6.2
transformers
accelerate
git+https://github.com/openai/whisper.git 
pyannote.audio
pydub
torchaudio==0.13.1
torchvision==0.14.1
